game as it is
